# Enforce HW_BUFFER_SIZE inside h1::dispatcher