@Dowwie 2020-06-03
> I hope you guys understand that when you read from socket you actually read from kernel memory.

with this given, what was missed?