@fafhrd91 2020-06-06
this patch fixes side-effect of the bug