@robjtede 2020-06-06
Sounds like you're saying there's a deeper issue here Nikolay? What might that be?