@Dowwie 2020-06-08
@actix/contributors Note this serious bug.  Also, the current fix adds unsafe, so would like to hear from others about concerns.  Large uploads are showing this behavior.   This issue requires escalating an emergency release once all are comfortable with the solution.