@robjtede 2020-06-08
I am investigating the root cause of this bug.