@vmalloc 2020-06-08
I'm not sure we're talking about the same thing, but if this is about backpressure when streaming payload to other destinations (e.g. streaming user payload to awc connections) I also think this is a serious issue. I tried implementing S3 uploads from an Actix backend and it ended up exposing my service to DoS when S3 was performing poorly.