@robo-corg 2020-06-08
It is possible a slow consumer might exacerbate this. For my own repro I just ran for_each on the stream of bytes. The problem that I saw was that `read_available` would would call `read` with the buffer as a `BufMut`. The `BufMut` for `BytesMut` can be resized arbitrarily which is presumably what the `AsyncRead` was doing. @fafhrd91 is probably right that this shouldn't be that big if we are reading whatever is available without blocking and so there is probably a bug at a lower level.

I was seeing this on loopback on a mac, and inside k8s on GKE. I might be able to build a minimal repro today if that is helpful. If we have minimal repro that will probably make debugging the underlying issue a lot easier.