@robo-corg 2020-06-08
Tested this under macos with instruments: https://github.com/robo-corg/actix-repro I might have time later today to investigate its behavior in linux. Running it in docker with top didn't seem to show it using very much memory.

On my machine it peaks at around 2gb and has multiple 1gb allocations coming from dispatch:

<img width="1202" alt="Screenshot 2020-06-08 13 35 33" src="https://user-images.githubusercontent.com/202000/84077956-f0e0c780-a98c-11ea-8af5-2b38277b0303.png">

After the patch with instruments peaks around 30mb no allocation by actix larger than 128kb (besides threads):
<img width="1740" alt="Screenshot 2020-06-08 13 45 50" src="https://user-images.githubusercontent.com/202000/84078825-65683600-a98e-11ea-9a3f-20cfe4d49f77.png">