@robjtede 2020-06-08
The problem is the loop in `read_available` which will allocate as much memory as it needs as long as the socket keeps polling as ready and providing data. Hard to say exactly what client conditions are necessary to trigger this but i've been able to reproduct very consitently using httpie but not curl.