@cetra3 2020-06-09
I agree with @robjtede that if we don't need to introduce a new `MaxBuf` (and the *unsafe* that goes along with it) we shouldn't.

But maybe 32k is not appropriate a value, and should be bumped up to something a bit larger?  Or even more of a refactor of the code here.

Is there also a way to easily test this behaviour that we can ensure there are no regressions in RAM usage?