@robjtede 2020-06-09
Some (admitedly non-exhaustive) performance numbers, appears to not be too harmful.


results in seconds for 1GB upload | original code | patched (32k) | patched (128k)
-- | -- | -- | --
httpie fresh | 1.951 | 1.951 | 4.087
httpie 2 | 1.981 | 3.994 | 1.575
httpie 3 | 1.995 | 1.959 | 1.785
  |   |   |  
curl fresh | 2.129 | 2.407 | 1.788
curl 2 | 1.717 | 1.646 | 2.185
curl 3 | 1.900 | 1.698 | 1.638