@Dowwie 2020-06-09
@robo-corg seems like all but the code that @robjtede identified could be cleared from the PR