@robo-corg 2020-06-09
I can update the patch to remove the MaxBuf. Probably run it through instruments and again and to verify.