@robo-corg 2020-06-09
You are right, seems to work exactly the same with MaxBuf gone so I removed it from this PR. It is a little disconcerting that the AsyncRead + BufMut contract allows it to enlarge the buffer but I guess that is fine. Same instruments output with no allocations going beyond 128k.

Do you want HW_BUFFER_SIZE modified? I could imagine that might be a nice thing to have as something configurable but maybe that should wait.