@robjtede 2020-06-09
Think we'll leave the buffer size alone for now, it didn't appear to have much of any effect in the large file tests.