@robo-corg 2020-06-17
I added an entry to `CHANGES.md`. I bump the version too if you want.