@adelbertc 2020-06-20
What is the protocol for getting this backported to actix-web 2.x (or actix-http 1.x)?