@robjtede 2020-09-11
@adelbertc just to follow up, we've decided not to backport any changes (for a multitude of reasons) but upgrading to v3 should be easy enough. There's a migration guide in the root of the repo that should guide you with anything that needs changing.